move verifier tests
