+++ crates/test-verifier/Cargo.toml
@@ -8,3 +8,13 @@ rust-version.workspace = true
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
+anyhow.workspace = true
+base64ct.workspace = true
+chrono.workspace = true
+credibil-did.workspace = true
+credibil-infosec.workspace = true
+credibil-vc = { workspace = true, features = ["issuer"] }
+ed25519-dalek = "2.1.1"
+serde.workspace = true
+serde_json.workspace = true
+uuid.workspace = true
