+++ crates/test-verifier/data/did-web.json
