+++ examples/verifier/Cargo.toml
@@ -8,13 +8,14 @@ workspace = true
 
 [dependencies]
 anyhow.workspace = true
-axum-extra.workspace = true
 axum.workspace = true
+axum-extra.workspace = true
 chrono.workspace = true
+credibil-vc = { workspace = true, features = ["verifier"] }
 serde.workspace = true
 serde_json.workspace = true
+test-verifier.workspace = true
 tokio.workspace = true
 tower-http.workspace = true
-tracing-subscriber.workspace = true
 tracing.workspace = true
-credibil-vc = { workspace = true, features = ["verifier"] }
+tracing-subscriber.workspace = true
