+++ rust-toolchain.toml
@@ -1,6 +1,6 @@
 [toolchain]
-channel = "stable"
-# channel = "nightly"
+# channel = "stable"
+channel = "nightly"
 components = ["clippy", "rust-src", "rustfmt"]
 targets = [
   "aarch64-apple-darwin",
