--- src/test_utils/sample.rs
@@ -1,9 +1,0 @@
-// use crate::openid::issuer::CredentialConfiguration;
-
-// pub fn credential_configuration() -> CredentialConfiguration {
-//     let issuer = super::store::issuance::IssuerStore::new()
-//         .get("http://credibil.io")
-//         .expect("should get issuer");
-
-//     issuer.credential_configurations_supported.get("EmployeeID_JWT").expect("should exist").clone()
-// }
