--- src/test_utils/store.rs
@@ -1,5 +1,0 @@
-pub mod issuance;
-pub mod keystore;
-pub mod presentation;
-pub mod resolver;
-pub mod state;
