--- src/test_utils/store/datasets.json
@@ -1,50 +1,0 @@
-{
-    "normal_user": {
-        "PHLEmployeeID": {
-            "configuration_id": "EmployeeID_JWT",
-            "claims": {
-                "given_name": "Normal",
-                "family_name": "Person",
-                "email": "normal.user@example.com",
-                "address": {
-                    "street_address": "123 Fake St",
-                    "locality": "Wellington"
-                }
-            },
-            "pending": false
-        },
-        "PHLDeveloper": {
-            "configuration_id": "Developer_JWT",
-            "claims": {
-                "given_name": "Normal",
-                "family_name": "Person",
-                "proficiency": "3"
-            },
-            "pending": false
-        },
-        "DriverLicence": {
-            "configuration_id": "org.iso.18013.5.1.mDL",
-            "claims": {
-                "given_name": "Normal",
-                "family_name": "Person",
-                "email": "normal.user@example.com"
-            },
-            "pending": false
-        }
-    },
-    "pending_user": {
-        "PHLEmployeeID": {
-            "configuration_id": "EmployeeID_JWT",
-            "claims": {
-                "given_name": "Pending",
-                "family_name": "Person",
-                "email": "pending.user@example.com",
-                "address": {
-                    "street_address": "123 Fake St",
-                    "locality": "Wellington"
-                }
-            },
-            "pending": true
-        }
-    }
-}
