--- src/test_utils/store/server.json
@@ -1,23 +1,0 @@
-{
-    "issuer": "http://credibil.io",
-    "authorization_endpoint": "/auth",
-    "token_endpoint": "/token",
-    "scopes_supported": [
-        "openid",
-        "credential"
-    ],
-    "response_types_supported": [
-        "code"
-    ],
-    "response_modes_supported": [
-        "query"
-    ],
-    "grant_types_supported": [
-        "authorization_code",
-        "urn:ietf:params:oauth:grant-type:pre-authorized_code"
-    ],
-    "code_challenge_methods_supported": [
-        "S256"
-    ],
-    "pre-authorized_grant_anonymous_access_supported": true
-}
