+++ src/verifier/request_object.rs
@@ -69,68 +69,3 @@ async fn process(
         request_object: RequestObjectType::Jwt(jwt_proof),
     })
 }
-
-#[cfg(test)]
-mod tests {
-    use chrono::Utc;
-    use credibil_infosec::jose::jws;
-    use insta::assert_yaml_snapshot as assert_snapshot;
-
-    use super::super::state::Expire;
-    use super::*;
-    use crate::core::Kind;
-    use crate::dif_exch::PresentationDefinition;
-    use crate::openid::verifier::{ClientIdScheme, RequestObject, ResponseType, Verifier};
-    use crate::test_utils::verifier::{Provider, VERIFIER_ID};
-    use crate::{test_utils, verify_key};
-
-    #[tokio::test]
-    async fn request_jwt() {
-        test_utils::init_tracer();
-
-        let provider = Provider::new();
-        let state_key = "ABCDEF123456";
-        let nonce = "1234567890";
-
-        let req_obj = RequestObject {
-            response_type: ResponseType::VpToken,
-            client_id: format!("{VERIFIER_ID}/post"),
-            state: Some(state_key.to_string()),
-            nonce: nonce.to_string(),
-            response_mode: Some("direct_post".into()),
-            response_uri: Some(format!("{VERIFIER_ID}/post")),
-            presentation_definition: Kind::Object(PresentationDefinition::default()),
-            client_id_scheme: Some(ClientIdScheme::RedirectUri),
-            client_metadata: Verifier::default(),
-
-            // TODO: populate missing RequestObject attributes
-            redirect_uri: None,
-            scope: None,
-        };
-
-        let state = State {
-            expires_at: Utc::now() + Expire::Request.duration(),
-            request_object: req_obj,
-        };
-        StateStore::put(&provider, &state_key, &state, state.expires_at)
-            .await
-            .expect("state exists");
-
-        let request = RequestObjectRequest {
-            client_id: VERIFIER_ID.to_string(),
-            id: state_key.to_string(),
-        };
-        let response = request_object(provider.clone(), &request).await.expect("response is valid");
-
-        let RequestObjectType::Jwt(jwt_enc) = &response.request_object else {
-            panic!("no JWT found in response");
-        };
-
-        let jwt: jws::Jwt<RequestObject> =
-            jws::decode(&jwt_enc, verify_key!(&provider)).await.expect("jwt is valid");
-        assert_snapshot!("response", jwt);
-
-        // request state should not exist
-        assert!(StateStore::get::<State>(&provider, state_key).await.is_ok());
-    }
-}
