+++ src/verifier/state.rs
@@ -7,12 +7,15 @@ use serde::{Deserialize, Serialize};
 
 use crate::openid::verifier::RequestObject;
 
+/// The duration for which a state item is valid.
 pub enum Expire {
+    /// The state item expires after the request is created.
     Request,
-    // Nonce,
 }
 
 impl Expire {
+    /// Returns the duration for which the state item is valid.
+    #[must_use]
     pub fn duration(&self) -> TimeDelta {
         match self {
             Self::Request => TimeDelta::try_minutes(5).unwrap_or_default(),
@@ -20,6 +23,7 @@ impl Expire {
     }
 }
 
+/// State is used by the library to persist request information between steps.
 #[derive(Builder, Clone, Debug, Default, Deserialize, Serialize)]
 pub struct State {
     /// The time this state item should expire.
@@ -30,10 +34,3 @@ pub struct State {
     /// Presentation Submission.
     pub request_object: RequestObject,
 }
-
-// impl State {
-//     /// Determines whether state has expired or not.
-//     pub fn expired(&self) -> bool {
-//         self.expires_at.signed_duration_since(Utc::now()).num_seconds() < 0
-//     }
-// }
