+++ tests/deferred.rs
@@ -1,24 +1,26 @@
 //! Tests for the `deferred` endpoint
 
+mod utils;
+
 use std::collections::HashMap;
 
 use assert_let_bind::assert_let;
 use chrono::Utc;
 use credibil_infosec::jose::JwsBuilder;
+use credibil_vc::issuer;
 use credibil_vc::issuer::state::{Authorized, Deferrance, Expire, Stage, State, Token};
 use credibil_vc::openid::issuer::{
     CredentialRequest, CredentialResponseType, DeferredCredentialRequest, ProofClaims,
 };
 use credibil_vc::openid::provider::StateStore;
 use credibil_vc::w3c_vc::proof::{self, Payload, Type, Verify};
-use credibil_vc::{issuer, snapshot};
 use insta::assert_yaml_snapshot as assert_snapshot;
 use serde_json::json;
 use test_issuer::{CLIENT_ID, CREDENTIAL_ISSUER, NORMAL_USER};
 
 #[tokio::test]
 async fn deferred_ok() {
-    // test_utils::init_tracer();
+    utils::init_tracer();
     snapshot!("");
     let provider = test_issuer::ProviderImpl::new();
