+++ tests/oauth_server.rs
@@ -1,13 +1,15 @@
 //! Tests for the OAuth server metadata.
 
+mod utils;
+
+use credibil_vc::issuer;
 use credibil_vc::openid::issuer::OAuthServerRequest;
-use credibil_vc::{issuer, snapshot};
 use insta::assert_yaml_snapshot as assert_snapshot;
 use test_issuer::CREDENTIAL_ISSUER;
 
 #[tokio::test]
 async fn metadata_ok() {
-    // test_utils::init_tracer();
+    utils::init_tracer();
     snapshot!("");
     let provider = test_issuer::ProviderImpl::new();
