+++ tests/par.rs
@@ -1,10 +1,12 @@
 //! Tests for Pushed Authorization Request endpoint
 
+mod utils;
+
 use base64ct::{Base64UrlUnpadded, Encoding};
+use credibil_vc::issuer;
 use credibil_vc::issuer::state::State;
 use credibil_vc::openid::issuer::{AuthorizationRequest, PushedAuthorizationRequest};
 use credibil_vc::openid::provider::StateStore;
-use credibil_vc::{issuer, snapshot};
 use insta::assert_yaml_snapshot as assert_snapshot;
 use serde_json::json;
 use sha2::{Digest, Sha256};
@@ -12,7 +14,7 @@ use test_issuer::{CLIENT_ID, CREDENTIAL_ISSUER, NORMAL_USER};
 
 #[tokio::test]
 async fn request() {
-    // test_utils::init_tracer();
+    utils::init_tracer();
     snapshot!("");
     let provider = test_issuer::ProviderImpl::new();
