+++ tests/utils.rs
@@ -1,9 +1,54 @@
 #![allow(missing_docs)]
+#![allow(dead_code)]
 
 use std::fmt::Display;
+use std::sync::Once;
 
 use rstest::fixture;
 use test_issuer::ProviderImpl;
+use tracing::Level;
+use tracing_subscriber::FmtSubscriber;
+
+// initalise tracing once for all tests
+static INIT: Once = Once::new();
+
+/// Configure Insta snapshot name.
+///
+/// # Example
+///
+/// ```rust,ignore
+/// snapshot!("issuer:immediate");
+///
+/// ...
+///
+/// assert_snapshot!("credential", vc, {
+///    ".validFrom" => "[validFrom]",
+///    ".credentialSubject" => insta::sorted_redaction()
+/// });
+/// ```
+///
+/// will result in a snapshot named `credential@issuer:immediate.snap`.
+#[macro_export]
+macro_rules! snapshot{
+    ($($expr:expr),*) => {
+        let mut settings = insta::Settings::clone_current();
+        settings.set_snapshot_suffix(format!($($expr,)*));
+        settings.set_prepend_module_to_snapshot(false);
+        let _guard = settings.bind_to_scope();
+    }
+}
+
+/// Initialise tracing for tests.
+///
+/// # Panics
+///
+/// Panics if the tracing subscriber cannot be set.
+pub fn init_tracer() {
+    INIT.call_once(|| {
+        let subscriber = FmtSubscriber::builder().with_max_level(Level::ERROR).finish();
+        tracing::subscriber::set_global_default(subscriber).expect("subscriber set");
+    });
+}
 
 #[fixture]
 pub fn provider() -> ProviderImpl {
@@ -12,7 +57,9 @@ pub fn provider() -> ProviderImpl {
 
 /// Issuance variants
 pub enum Issuance {
+    /// Immediate issuance
     Immediate,
+    /// Deferred issuance
     Deferred,
 }
