+++ tests/verifier_metadata.rs
@@ -1,0 +1,22 @@
+//! Tests for the verifier metadata endpoint.
+
+mod utils;
+
+// use providers::wallet_provider::holder_provider::CLIENT_ID;
+use credibil_vc::openid::verifier::MetadataRequest;
+use credibil_vc::verifier;
+use insta::assert_yaml_snapshot as assert_snapshot;
+
+#[tokio::test]
+async fn metadata_ok() {
+    utils::init_tracer();
+    let provider = test_verifier::ProviderImpl::new();
+
+    let request = MetadataRequest {
+        client_id: "http://localhost:8080".into(),
+    };
+    let response = verifier::metadata(provider, &request).await.expect("response is ok");
+    assert_snapshot!("response", response, {
+        ".vp_formats" => insta::sorted_redaction()
+    });
+}
