+++ tests/wallet_initiated.rs
@@ -4,7 +4,6 @@ mod utils;
 mod wallet;
 
 use credibil_vc::issuer::{Format, ProfileW3c};
-use credibil_vc::{snapshot, test_utils};
 use rstest::rstest;
 use test_issuer::ProviderImpl;
 use utils::{Issuance, provider};
@@ -13,7 +12,7 @@ use utils::{Issuance, provider};
 #[case(Issuance::Immediate)]
 #[case(Issuance::Deferred)]
 async fn issuance(provider: ProviderImpl, #[case] issue: Issuance) {
-    test_utils::init_tracer();
+    utils::init_tracer();
     snapshot!("wallet:{issue}");
 
     let wallet = wallet::Wallet {
